# CUDA: use tensor cores for MMQ