This PR aims to add int8 tensor core support for mul_mat_q kernels (legacy quants only for now). The supported hardware will be Turing or newer. So far there is only a prototype for q8_0 which on its own is still slower than FP16 cuBLAS but faster for end-to-end performance because it needs less data conversion. Current performance:

| GPU      | Model         |   Microbatch size | Test   |   t/s master |   t/s PR |   Speedup |
|:---------|:--------------|------------------:|:-------|-------------:|---------------:|----------:|
| RTX 4090 | llama 8B Q8_0 |                64 | pp4096 |      1688.32 |        3107.57 |      1.84 |
| RTX 4090 | llama 8B Q8_0 |               128 | pp4096 |      3208.06 |        5515.37 |      1.72 |
| RTX 4090 | llama 8B Q8_0 |               256 | pp4096 |      5297.47 |        8062.99 |      1.52 |
| RTX 4090 | llama 8B Q8_0 |               512 | pp4096 |      7095.77 |        9458.58 |      1.33 |
| RTX 4090 | llama 8B Q8_0 |              1024 | pp4096 |      8322.32 |        9533.70 |      1.15 |
| RTX 4090 | llama 8B Q8_0 |              2048 | pp4096 |      8594.14 |        9131.38 |      1.06 |
| RTX 4090 | llama 8B Q8_0 |              4096 | pp4096 |      8593.86 |        9127.01 |      1.06 |

As of right now this PR **must** be compiled with `LLAMA_CUDA_FORCE_MMQ`. `scripts/copare_llama_bench.py` needs the fix added by https://github.com/ggerganov/llama.cpp/pull/7673 .

The way to make int8 tensor cores work is to write PTX code (the CUDA equivalent of assembly) because with the "high level" WMMA interface you do not have a defined memory layout which makes it impossible to correctly apply the scales of ggml quantized data blocks. I plan to wrap the PTX code in simple CUDA functions in order to hopefully make it easier to understand what it does.