@slaren 2024-05-31
| model                          |         fa |          test |     master cuBLAS t/s | PR MMQ t/s | speedup |
| ------------------------------ | ---------: | ------------: | ---------------: |---------------: | -------:|
| llama 7B Q8_0                  |          1 |          pp32 |    462.62 ± 0.62 |  1179.08 ± 5.90 |  2.548 |
| llama 7B Q8_0                  |          1 |          pp64 |   1406.62 ± 3.72 |  2249.44 ± 9.03 |  1.599 |
| llama 7B Q8_0                  |          1 |         pp128 |   2475.62 ± 3.91 | 4036.57 ± 14.50 |  1.630 |
| llama 7B Q8_0                  |          1 |         pp256 |   3921.04 ± 7.08 |  5005.20 ± 7.61 |  1.276 |
| llama 7B Q8_0                  |          1 |         pp512 |   4816.36 ± 8.50 | 5237.32 ± 11.06 |  1.087 |
| llama 7B Q8_0                  |          1 |        pp1024 |   5702.84 ± 4.58 |  5483.00 ± 9.20 |  0.961 |
| llama 7B Q8_0                  |          1 |        pp2048 |   5768.32 ± 7.25 | 5367.63 ± 13.85 |  0.930 |
| llama 7B Q8_0                  |          1 |        pp4096 |   5300.56 ± 5.46 |  4818.36 ± 3.07 |  0.909 |


`llama-bench` still limits the batch size to the value of `n_batch`, even with it is smaller than `n_ubatch`. The default `n_batch` is 2048, so this may be why you are seeing the same performance for 2048 and 4096. It should probably be smarter about that, or at least use the value after it is adjusted by llama.cpp.