@JohannesGaessler 2024-06-08
I implemented support for q4_0, q4_1, q5_0, q5_1, and q8_0 based on https://github.com/ggerganov/llama.cpp/pull/7824 . The performance currently looks like this:

<details>
<summary>Vs. master MMQ</summary>

| GPU        | Model           |     Microbatch size | Test     |     t/s master |     t/s cuda-ptx-mma-7 |     Speedup |
| :--------- | :-------------- | ------------------: | :------- | -------------: | ---------------------: | ----------: |
| RTX 4090   | llama 8B Q4_0   |                  16 | pp512    |        1601.20 |                1937.21 |        1.21 |
| RTX 4090   | llama 8B Q4_0   |                  32 | pp512    |        2620.80 |                3427.61 |        1.31 |
| RTX 4090   | llama 8B Q4_0   |                  64 | pp512    |        3704.68 |                5420.06 |        1.46 |
| RTX 4090   | llama 8B Q4_0   |                 128 | pp512    |        4991.25 |                7443.66 |        1.49 |
| RTX 4090   | llama 8B Q4_0   |                 256 | pp512    |        5833.07 |                9601.53 |        1.65 |
| RTX 4090   | llama 8B Q4_0   |                 512 | pp512    |        6279.91 |               10358.75 |        1.65 |
| RTX 4090   | llama 8B Q4_1   |                  16 | pp512    |        1558.18 |                1888.20 |        1.21 |
| RTX 4090   | llama 8B Q4_1   |                  32 | pp512    |        2580.88 |                3252.02 |        1.26 |
| RTX 4090   | llama 8B Q4_1   |                  64 | pp512    |        3685.14 |                5778.21 |        1.57 |
| RTX 4090   | llama 8B Q4_1   |                 128 | pp512    |        4809.24 |                6765.41 |        1.41 |
| RTX 4090   | llama 8B Q4_1   |                 256 | pp512    |        5606.55 |                9152.41 |        1.63 |
| RTX 4090   | llama 8B Q4_1   |                 512 | pp512    |        5853.33 |                9795.29 |        1.67 |
| RTX 4090   | llama 8B Q5_0   |                  16 | pp512    |        1252.35 |                1520.90 |        1.21 |
| RTX 4090   | llama 8B Q5_0   |                  32 | pp512    |        1988.05 |                2564.28 |        1.29 |
| RTX 4090   | llama 8B Q5_0   |                  64 | pp512    |        2916.94 |                4413.28 |        1.51 |
| RTX 4090   | llama 8B Q5_0   |                 128 | pp512    |        4224.09 |                6479.58 |        1.53 |
| RTX 4090   | llama 8B Q5_0   |                 256 | pp512    |        5138.20 |                8278.94 |        1.61 |
| RTX 4090   | llama 8B Q5_0   |                 512 | pp512    |        5773.74 |                9434.57 |        1.63 |
| RTX 4090   | llama 8B Q5_1   |                  16 | pp512    |        1300.38 |                1669.12 |        1.28 |
| RTX 4090   | llama 8B Q5_1   |                  32 | pp512    |        2157.43 |                2692.43 |        1.25 |
| RTX 4090   | llama 8B Q5_1   |                  64 | pp512    |        2989.86 |                4581.81 |        1.53 |
| RTX 4090   | llama 8B Q5_1   |                 128 | pp512    |        4229.24 |                6192.37 |        1.46 |
| RTX 4090   | llama 8B Q5_1   |                 256 | pp512    |        5015.76 |                8026.92 |        1.60 |
| RTX 4090   | llama 8B Q5_1   |                 512 | pp512    |        5542.11 |                9070.31 |        1.64 |
| RTX 4090   | llama 8B Q8_0   |                  16 | pp512    |        1035.64 |                1134.04 |        1.10 |
| RTX 4090   | llama 8B Q8_0   |                  32 | pp512    |        1793.27 |                2216.12 |        1.24 |
| RTX 4090   | llama 8B Q8_0   |                  64 | pp512    |        2980.14 |                4140.20 |        1.39 |
| RTX 4090   | llama 8B Q8_0   |                 128 | pp512    |        4323.77 |                6573.77 |        1.52 |
| RTX 4090   | llama 8B Q8_0   |                 256 | pp512    |        5308.85 |                8961.02 |        1.69 |
| RTX 4090   | llama 8B Q8_0   |                 512 | pp512    |        5888.12 |               10139.99 |        1.72 |
| RTX 3090   | llama 8B Q4_0   |                  16 | pp512    |         897.63 |                1178.88 |        1.31 |
| RTX 3090   | llama 8B Q4_0   |                  32 | pp512    |        1194.11 |                1752.72 |        1.47 |
| RTX 3090   | llama 8B Q4_0   |                  64 | pp512    |        1655.52 |                2470.20 |        1.49 |
| RTX 3090   | llama 8B Q4_0   |                 128 | pp512    |        1938.21 |                3227.85 |        1.67 |
| RTX 3090   | llama 8B Q4_0   |                 256 | pp512    |        2374.01 |                3892.61 |        1.64 |
| RTX 3090   | llama 8B Q4_0   |                 512 | pp512    |        2483.79 |                4148.35 |        1.67 |
| RTX 3090   | llama 8B Q4_1   |                  16 | pp512    |         883.36 |                1301.24 |        1.47 |
| RTX 3090   | llama 8B Q4_1   |                  32 | pp512    |        1180.58 |                1847.37 |        1.56 |
| RTX 3090   | llama 8B Q4_1   |                  64 | pp512    |        1544.92 |                2390.60 |        1.55 |
| RTX 3090   | llama 8B Q4_1   |                 128 | pp512    |        1851.64 |                3130.67 |        1.69 |
| RTX 3090   | llama 8B Q4_1   |                 256 | pp512    |        2238.07 |                3646.03 |        1.63 |
| RTX 3090   | llama 8B Q4_1   |                 512 | pp512    |        2363.05 |                4068.86 |        1.72 |
| RTX 3090   | llama 8B Q5_0   |                  16 | pp512    |         618.30 |                 850.13 |        1.37 |
| RTX 3090   | llama 8B Q5_0   |                  32 | pp512    |         879.04 |                1367.69 |        1.56 |
| RTX 3090   | llama 8B Q5_0   |                  64 | pp512    |        1278.90 |                1951.02 |        1.53 |
| RTX 3090   | llama 8B Q5_0   |                 128 | pp512    |        1683.43 |                2743.75 |        1.63 |
| RTX 3090   | llama 8B Q5_0   |                 256 | pp512    |        2144.71 |                3518.48 |        1.64 |
| RTX 3090   | llama 8B Q5_0   |                 512 | pp512    |        2256.40 |                3654.78 |        1.62 |
| RTX 3090   | llama 8B Q5_1   |                  16 | pp512    |         686.36 |                1014.38 |        1.48 |
| RTX 3090   | llama 8B Q5_1   |                  32 | pp512    |         894.08 |                1530.24 |        1.71 |
| RTX 3090   | llama 8B Q5_1   |                  64 | pp512    |        1298.23 |                1986.79 |        1.53 |
| RTX 3090   | llama 8B Q5_1   |                 128 | pp512    |        1638.68 |                2667.84 |        1.63 |
| RTX 3090   | llama 8B Q5_1   |                 256 | pp512    |        2053.43 |                3355.29 |        1.63 |
| RTX 3090   | llama 8B Q5_1   |                 512 | pp512    |        2190.26 |                3493.64 |        1.60 |
| RTX 3090   | llama 8B Q8_0   |                  16 | pp512    |         575.23 |                 825.82 |        1.44 |
| RTX 3090   | llama 8B Q8_0   |                  32 | pp512    |         934.07 |                1437.37 |        1.54 |
| RTX 3090   | llama 8B Q8_0   |                  64 | pp512    |        1357.47 |                2242.88 |        1.65 |
| RTX 3090   | llama 8B Q8_0   |                 128 | pp512    |        1760.07 |                3132.53 |        1.78 |
| RTX 3090   | llama 8B Q8_0   |                 256 | pp512    |        2168.84 |                3958.13 |        1.82 |
| RTX 3090   | llama 8B Q8_0   |                 512 | pp512    |        2359.58 |                4185.76 |        1.77 |

</details>

<details>
<summary>Vs. master cuBLAS</summary>

| GPU        | Model           |     Microbatch size | Test     |     t/s master |     t/s cuda-ptx-mma-7 |     Speedup |
| :--------- | :-------------- | ------------------: | :------- | -------------: | ---------------------: | ----------: |
| RTX 4090   | llama 8B Q4_0   |                  16 | pp512    |        1596.75 |                1937.21 |        1.21 |
| RTX 4090   | llama 8B Q4_0   |                  32 | pp512    |        2621.55 |                3427.61 |        1.31 |
| RTX 4090   | llama 8B Q4_0   |                  64 | pp512    |        3704.48 |                5420.06 |        1.46 |
| RTX 4090   | llama 8B Q4_0   |                 128 | pp512    |        3578.31 |                7443.66 |        2.08 |
| RTX 4090   | llama 8B Q4_0   |                 256 | pp512    |        5838.52 |                9601.53 |        1.64 |
| RTX 4090   | llama 8B Q4_0   |                 512 | pp512    |        7776.11 |               10358.75 |        1.33 |
| RTX 4090   | llama 8B Q4_1   |                  16 | pp512    |        1558.45 |                1888.20 |        1.21 |
| RTX 4090   | llama 8B Q4_1   |                  32 | pp512    |        2578.79 |                3252.02 |        1.26 |
| RTX 4090   | llama 8B Q4_1   |                  64 | pp512    |        3683.87 |                5778.21 |        1.57 |
| RTX 4090   | llama 8B Q4_1   |                 128 | pp512    |        3540.69 |                6765.41 |        1.91 |
| RTX 4090   | llama 8B Q4_1   |                 256 | pp512    |        5787.13 |                9152.41 |        1.58 |
| RTX 4090   | llama 8B Q4_1   |                 512 | pp512    |        7671.16 |                9795.29 |        1.28 |
| RTX 4090   | llama 8B Q5_0   |                  16 | pp512    |        1251.55 |                1520.90 |        1.22 |
| RTX 4090   | llama 8B Q5_0   |                  32 | pp512    |        1987.14 |                2564.28 |        1.29 |
| RTX 4090   | llama 8B Q5_0   |                  64 | pp512    |        2914.51 |                4413.28 |        1.51 |
| RTX 4090   | llama 8B Q5_0   |                 128 | pp512    |        3477.98 |                6479.58 |        1.86 |
| RTX 4090   | llama 8B Q5_0   |                 256 | pp512    |        5726.53 |                8278.94 |        1.45 |
| RTX 4090   | llama 8B Q5_0   |                 512 | pp512    |        7677.37 |                9434.57 |        1.23 |
| RTX 4090   | llama 8B Q5_1   |                  16 | pp512    |        1296.52 |                1669.12 |        1.29 |
| RTX 4090   | llama 8B Q5_1   |                  32 | pp512    |        2152.98 |                2692.43 |        1.25 |
| RTX 4090   | llama 8B Q5_1   |                  64 | pp512    |        2983.69 |                4581.81 |        1.54 |
| RTX 4090   | llama 8B Q5_1   |                 128 | pp512    |        3490.47 |                6192.37 |        1.77 |
| RTX 4090   | llama 8B Q5_1   |                 256 | pp512    |        5703.59 |                8026.92 |        1.41 |
| RTX 4090   | llama 8B Q5_1   |                 512 | pp512    |        7632.39 |                9070.31 |        1.19 |
| RTX 4090   | llama 8B Q8_0   |                  16 | pp512    |        1035.42 |                1134.04 |        1.10 |
| RTX 4090   | llama 8B Q8_0   |                  32 | pp512    |        1793.94 |                2216.12 |        1.24 |
| RTX 4090   | llama 8B Q8_0   |                  64 | pp512    |        2981.81 |                4140.20 |        1.39 |
| RTX 4090   | llama 8B Q8_0   |                 128 | pp512    |        3376.19 |                6573.77 |        1.95 |
| RTX 4090   | llama 8B Q8_0   |                 256 | pp512    |        5575.87 |                8961.02 |        1.61 |
| RTX 4090   | llama 8B Q8_0   |                 512 | pp512    |        7511.04 |               10139.99 |        1.35 |
| RTX 3090   | llama 8B Q4_0   |                  16 | pp512    |         899.84 |                1178.88 |        1.31 |
| RTX 3090   | llama 8B Q4_0   |                  32 | pp512    |        1188.91 |                1752.72 |        1.47 |
| RTX 3090   | llama 8B Q4_0   |                  64 | pp512    |        1653.41 |                2470.20 |        1.49 |
| RTX 3090   | llama 8B Q4_0   |                 128 | pp512    |        2347.86 |                3227.85 |        1.37 |
| RTX 3090   | llama 8B Q4_0   |                 256 | pp512    |        3516.58 |                3892.61 |        1.11 |
| RTX 3090   | llama 8B Q4_0   |                 512 | pp512    |        4094.19 |                4148.35 |        1.01 |
| RTX 3090   | llama 8B Q4_1   |                  16 | pp512    |         877.33 |                1301.24 |        1.48 |
| RTX 3090   | llama 8B Q4_1   |                  32 | pp512    |        1169.76 |                1847.37 |        1.58 |
| RTX 3090   | llama 8B Q4_1   |                  64 | pp512    |        1517.09 |                2390.60 |        1.58 |
| RTX 3090   | llama 8B Q4_1   |                 128 | pp512    |        2323.90 |                3130.67 |        1.35 |
| RTX 3090   | llama 8B Q4_1   |                 256 | pp512    |        3469.03 |                3646.03 |        1.05 |
| RTX 3090   | llama 8B Q4_1   |                 512 | pp512    |        4024.70 |                4068.86 |        1.01 |
| RTX 3090   | llama 8B Q5_0   |                  16 | pp512    |         621.64 |                 850.13 |        1.37 |
| RTX 3090   | llama 8B Q5_0   |                  32 | pp512    |         873.34 |                1367.69 |        1.57 |
| RTX 3090   | llama 8B Q5_0   |                  64 | pp512    |        1257.98 |                1951.02 |        1.55 |
| RTX 3090   | llama 8B Q5_0   |                 128 | pp512    |        2182.46 |                2743.75 |        1.26 |
| RTX 3090   | llama 8B Q5_0   |                 256 | pp512    |        3286.80 |                3518.48 |        1.07 |
| RTX 3090   | llama 8B Q5_0   |                 512 | pp512    |        3968.35 |                3654.78 |        0.92 |
| RTX 3090   | llama 8B Q5_1   |                  16 | pp512    |         681.66 |                1014.38 |        1.49 |
| RTX 3090   | llama 8B Q5_1   |                  32 | pp512    |         876.71 |                1530.24 |        1.75 |
| RTX 3090   | llama 8B Q5_1   |                  64 | pp512    |        1275.26 |                1986.79 |        1.56 |
| RTX 3090   | llama 8B Q5_1   |                 128 | pp512    |        2161.81 |                2667.84 |        1.23 |
| RTX 3090   | llama 8B Q5_1   |                 256 | pp512    |        3264.62 |                3355.29 |        1.03 |
| RTX 3090   | llama 8B Q5_1   |                 512 | pp512    |        3923.75 |                3493.64 |        0.89 |
| RTX 3090   | llama 8B Q8_0   |                  16 | pp512    |         571.46 |                 825.82 |        1.45 |
| RTX 3090   | llama 8B Q8_0   |                  32 | pp512    |         918.53 |                1437.37 |        1.56 |
| RTX 3090   | llama 8B Q8_0   |                  64 | pp512    |        1335.22 |                2242.88 |        1.68 |
| RTX 3090   | llama 8B Q8_0   |                 128 | pp512    |        2201.87 |                3132.53 |        1.42 |
| RTX 3090   | llama 8B Q8_0   |                 256 | pp512    |        3327.89 |                3958.13 |        1.19 |
| RTX 3090   | llama 8B Q8_0   |                 512 | pp512    |        3979.69 |                4185.76 |        1.05 |

</details>

My immediate next goals will be in order to add support for k-quants, optimize the performance, and to refactor and simplify the code.